{image,package}/python: harden PGO stage

+++ package/python/template/bench.sh
@@ -5,7 +5,7 @@ set -f
 : "${DEB_SRC_TOPDIR:?}" "${DEB_PGO_FROM_BUILD:?}" "${DEB_PGO_FROM_PKG:?}"
 
 while [ "${DEB_PGO_REUSE}" = yes ] ; do
-	[ -d "/${DEB_PGO_FROM_PKG}" ] || break
+	[ -d "/${DEB_PGO_FROM_PKG}" ] || exit 1
 
 	find ./ -name '*.gcda' -printf '%P\0' \
 	| sort -zuV | xargs -0r rm -fv
@@ -30,12 +30,12 @@ end_script() {
 	| sort -zuV | tar --null -T - -cf - \
 	| tar -C "${DEB_PGO_FROM_BUILD}" -xvf -
 
-	exit 0
+	exit "${1:-0}"
 }
 
 end_if_level() {
-	[ "${DEB_PGO_LEVEL}" = "$1" ] || return
-	end_script
+	[ "${DEB_PGO_LEVEL}" = "$1" ] || return 0
+	end_script 0
 }
 
 unset CONTAINER_PYTHON_COMPAT
@@ -45,7 +45,7 @@ unset CONTAINER_PYTHON_COMPAT
 do_python_tests() {
 	# run in subshell
 	flush_pycache
-	( export CONTAINER_PYTHON_COMPAT=1 ; set -xv ; "$@" ; ) || end_script
+	( export CONTAINER_PYTHON_COMPAT=1 ; set -xv ; "$@" ; ) || end_script 1
 	flush_pycache
 }
 
@@ -58,14 +58,18 @@ end_if_level 1
 python_bin=$(readlink -f "$1")
 python_wrap=$(readlink -f ./runpython.sh)
 
+do_pip_install() {
+	"${python_wrap}" -m pip install "$@" || end_script 1
+}
+
 ## pyperformance
 
-"${python_wrap}" -m pip install \
+do_pip_install \
   "${DEB_SRC_TOPDIR}/pip-pyperformance"
 
 do_pyperf() {
 	flush_pycache
-	"${python_wrap}" -m pyperformance "$@" || end_script
+	"${python_wrap}" -m pyperformance "$@" || end_script 1
 	flush_pycache
 }
 
@@ -75,12 +79,13 @@ end_if_level 2
 
 ## asv-based tests/benchmarks
 
-"${python_wrap}" -m pip install \
+do_pip_install \
   'asv~=0.5.1' \
   'virtualenv~=20.24.0' \
 
 
 do_asv() { "${python_wrap}" -m asv "$@" ; }
+# TODO: fix all asv-based tests and stop ignoring errors :)
 do_asv_at() {
 	cd "$1"
 	do_asv machine --yes >&2
@@ -88,7 +93,7 @@ do_asv_at() {
 	date -R >&2
 	echo >&2
 	flush_pycache
-	do_asv run --quick --parallel 1 --no-pull --dry-run --show-stderr --environment "existing:${python_bin}" || end_script
+	do_asv run --quick --parallel 1 --no-pull --dry-run --show-stderr --environment "existing:${python_bin}"
 	flush_pycache
 	echo >&2
 	date -R >&2
@@ -98,7 +103,7 @@ do_asv_at() {
 
 ## asv: numpy
 
-"${python_wrap}" -m pip install \
+do_pip_install \
   "numpy==${NUMPY_VERSION}" \
   'cython~=0.29.36' \
 
@@ -108,7 +113,7 @@ end_if_level 3
 
 ## asv: dask/distributed
 
-"${python_wrap}" -m pip install \
+do_pip_install \
   "dask==${DASK_VERSION}" \
   'cython~=0.29.36' \
   'numpy~=1.24.0' \
@@ -121,7 +126,7 @@ end_if_level 3
 do_asv_at "${DEB_SRC_TOPDIR}/pip-dask-benchmarks/dask"
 end_if_level 4
 
-"${python_wrap}" -m pip install \
+do_pip_install \
   "distributed==${DASK_VERSION}" \
   'cython~=0.29.36' \
   'lz4~=4.3.2' \
@@ -133,7 +138,7 @@ end_if_level 5
 
 ## asv: pandas
 
-"${python_wrap}" -m pip install \
+do_pip_install \
   "pandas==${PANDAS_VERSION}" \
   'cython~=0.29.36' \
   'jinja2~=3.1.2' \
